feat: combine pager components
