+++ src/pages/Lend/LoanChannelCategoryControl.vue
@@ -138,7 +138,13 @@
 					/>
 					<kv-loading-overlay v-if="loading" />
 				</div>
-				<kv-pagination v-if="totalCount > 0" :total="totalCount" :limit="limit" @page-change="pageChange" />
+				<kv-pagination
+					v-if="totalCount > 0"
+					:total="totalCount"
+					:limit="limit"
+					:offset="offset"
+					@page-changed="pageChange"
+				/>
 				<div v-if="totalCount > 0" class="loan-count tw-text-tertiary">
 					{{ totalCount }} loans
 				</div>
@@ -471,7 +477,7 @@ export default {
 			const loansOutOfRange = loansArrayLength === 0 && pageQueryParam;
 			if (loansOutOfRange) {
 				this.$showTipMsg(`There are currently ${this.lastLoanPage} pages of results. We’ve loaded the last page for you.`); // eslint-disable-line max-len
-				this.pageChange(this.lastLoanPage);
+				this.pageChange({ pageOffset: loansPerPage * (this.lastLoanPage - 1) });
 			}
 		},
 		updateLoanReservation(id) {
@@ -485,9 +491,8 @@ export default {
 				})
 			);
 		},
-		pageChange(number) {
-			const offset = loansPerPage * (number - 1);
-			this.offset = offset;
+		pageChange({ pageOffset }) {
+			this.offset = pageOffset;
 			this.pushChangesToUrl();
 		},
 		updateFromParams(query) {
