+++ src/pages/Lend/LoanChannelCategoryRecommendedByLenders.vue
@@ -39,7 +39,13 @@
 						loan-card-type="GridLoanCard"
 					/>
 				</div>
-				<kv-pagination v-if="totalCount > 0" :total="totalCount" :limit="limit" @page-change="pageChange" />
+				<kv-pagination
+					v-if="totalCount > 0"
+					:total="totalCount"
+					:limit="limit"
+					:offset="offset"
+					@page-changed="pageChange"
+				/>
 				<div v-if="totalCount > 0" class="loan-count tw-text-tertiary">
 					{{ totalCount }} loans
 				</div>
@@ -307,6 +313,10 @@ export default {
 		this.targetedLoanChannelID = getTargetedChannel(targetedLoanChannelURL, allChannelsData);
 		// extract query
 		this.pageQuery = _get(this.$route, 'query');
+
+		// Ensure page offset gets set before loading cached data
+		this.updateFromParams(this.pageQuery);
+
 		// Read the page data from the cache
 		let baseData = {};
 		try {
@@ -350,7 +360,7 @@ export default {
 			const loansOutOfRange = loansArrayLength === 0 && pageQueryParam;
 			if (loansOutOfRange) {
 				this.$showTipMsg(`There are currently ${this.lastLoanPage} pages of results. We’ve loaded the last page for you.`); // eslint-disable-line max-len
-				this.pageChange(this.lastLoanPage);
+				this.pageChange({ pageOffset: loansPerPage * (this.lastLoanPage - 1) });
 			}
 		},
 		updateLoanReservation(id) {
@@ -364,9 +374,8 @@ export default {
 				})
 			);
 		},
-		pageChange(number) {
-			const offset = loansPerPage * (number - 1);
-			this.offset = offset;
+		pageChange({ pageOffset }) {
+			this.offset = pageOffset;
 			this.pushChangesToUrl();
 		},
 		updateFromParams(query) {
