+++ test/unit/specs/components/Kv/KvPagination.spec.js
@@ -1,16 +1,16 @@
 import { render } from '@testing-library/vue';
 import userEvent from '@testing-library/user-event';
-import KvPager from '@/components/Kv/KvPager';
+import KvPagination from '@/components/Kv/KvPagination';
 
 global.scrollTo = jest.fn();
 
-describe('KvPager', () => {
+describe('KvPagination', () => {
 	afterEach(jest.clearAllMocks);
 
 	it('should render arrows disabled by default', async () => {
 		const user = userEvent.setup();
 
-		const { getByLabelText, emitted } = render(KvPager, { props: { limit: 10, total: 0 } });
+		const { getByLabelText, emitted } = render(KvPagination, { props: { limit: 10, total: 0 } });
 
 		await user.click(getByLabelText('Previous page'));
 		await user.click(getByLabelText('Next page'));
@@ -19,13 +19,13 @@ describe('KvPager', () => {
 	});
 
 	it('should render aria current page label', () => {
-		const { getByText } = render(KvPager, { props: { limit: 10, total: 30 } });
+		const { getByText } = render(KvPagination, { props: { limit: 10, total: 30 } });
 
 		getByText("You're on page");
 	});
 
 	it('should render fewer pages', () => {
-		const { getByLabelText } = render(KvPager, { props: { limit: 10, total: 30 } });
+		const { getByLabelText } = render(KvPagination, { props: { limit: 10, total: 30 } });
 
 		getByLabelText('Page 1');
 		getByLabelText('Page 2');
@@ -33,7 +33,7 @@ describe('KvPager', () => {
 	});
 
 	it('should render more pages', () => {
-		const { getByLabelText } = render(KvPager, { props: { limit: 10, total: 1000 } });
+		const { getByLabelText } = render(KvPagination, { props: { limit: 10, total: 1000 } });
 
 		getByLabelText('Page 1');
 		getByLabelText('Page 2');
@@ -43,7 +43,7 @@ describe('KvPager', () => {
 	});
 
 	it('should render fourth selected', () => {
-		const { getByLabelText } = render(KvPager, { props: { limit: 10, total: 1000, offset: 30 } });
+		const { getByLabelText } = render(KvPagination, { props: { limit: 10, total: 1000, offset: 30 } });
 
 		getByLabelText('Page 1');
 		getByLabelText('Page 3');
@@ -53,7 +53,7 @@ describe('KvPager', () => {
 	});
 
 	it('should render last selected', () => {
-		const { getByLabelText } = render(KvPager, { props: { limit: 10, total: 1000, offset: 990 } });
+		const { getByLabelText } = render(KvPagination, { props: { limit: 10, total: 1000, offset: 990 } });
 
 		getByLabelText('Page 1');
 		getByLabelText('Page 97');
@@ -63,7 +63,7 @@ describe('KvPager', () => {
 	});
 
 	it('should render fourth to last last selected', () => {
-		const { getByLabelText } = render(KvPager, { props: { limit: 10, total: 1000, offset: 960 } });
+		const { getByLabelText } = render(KvPagination, { props: { limit: 10, total: 1000, offset: 960 } });
 
 		getByLabelText('Page 1');
 		getByLabelText('Page 96');
@@ -73,7 +73,7 @@ describe('KvPager', () => {
 	});
 
 	it('should render more extra pages', () => {
-		const { getByLabelText } = render(KvPager, { props: { limit: 10, total: 1000, extraPages: 6 } });
+		const { getByLabelText } = render(KvPagination, { props: { limit: 10, total: 1000, extraPages: 6 } });
 
 		getByLabelText('Page 1');
 		getByLabelText('Page 2');
@@ -88,7 +88,7 @@ describe('KvPager', () => {
 	it('should emit page click', async () => {
 		const user = userEvent.setup();
 
-		const { getByLabelText, emitted } = render(KvPager, { props: { limit: 10, total: 1000 } });
+		const { getByLabelText, emitted } = render(KvPagination, { props: { limit: 10, total: 1000 } });
 
 		await user.click(getByLabelText('Page 2'));
 
@@ -99,7 +99,7 @@ describe('KvPager', () => {
 	it('should not emit current page click', async () => {
 		const user = userEvent.setup();
 
-		const { getByLabelText, emitted } = render(KvPager, { props: { limit: 10, total: 1000 } });
+		const { getByLabelText, emitted } = render(KvPagination, { props: { limit: 10, total: 1000 } });
 
 		await user.click(getByLabelText('Page 1'));
 
@@ -109,7 +109,7 @@ describe('KvPager', () => {
 	it('should emit previous click', async () => {
 		const user = userEvent.setup();
 
-		const { getByLabelText, emitted } = render(KvPager, { props: { limit: 10, total: 1000, offset: 10 } });
+		const { getByLabelText, emitted } = render(KvPagination, { props: { limit: 10, total: 1000, offset: 10 } });
 
 		await user.click(getByLabelText('Previous page'));
 
@@ -120,7 +120,7 @@ describe('KvPager', () => {
 	it('should emit next click', async () => {
 		const user = userEvent.setup();
 
-		const { getByLabelText, emitted } = render(KvPager, { props: { limit: 10, total: 1000 } });
+		const { getByLabelText, emitted } = render(KvPagination, { props: { limit: 10, total: 1000 } });
 
 		await user.click(getByLabelText('Next page'));
